Add files via upload
It's the improved version of the CCAR which is now capable to handle partial funding issues and enables users to manage a flexible bond portfolio by adding some more new bonds(loans) into the existing portfolio at the end of each season. The model is still built on a quarter base.